requirements.txt
websockets
pillow
aiohttp
flask
requests